+++ src/plonemeeting/portal/core/content/institution.py
@@ -251,19 +251,26 @@ class Institution(Container):
     def fetch_delib_categories(self):
         categories = []
         if self.plonemeeting_url and self.meeting_config_id and self.username and self.password:
-            delib_config_category_field = CATEGORY_IA_DELIB_FIELDS_MAPPING_EXTRA_INCLUDE[self.delib_category_field]
+            delib_config_category_field = CATEGORY_IA_DELIB_FIELDS_MAPPING_EXTRA_INCLUDE[
+                self.delib_category_field]
             url = get_api_url_for_categories(self, delib_config_category_field)
             if url:
                 logger.info("Fetching delib categories for {} [Start]".format(self.title))
                 response = requests.get(
                     url, auth=(self.username, self.password), headers=API_HEADERS
                 )
-                delib_config_category_field = CATEGORY_IA_DELIB_FIELDS_MAPPING_EXTRA_INCLUDE[self.delib_category_field]
-                json = response.json()
-                cat_json = json["extra_include_{categories}".format(categories=delib_config_category_field)]
+                if response.status_code in (200, 201):
+                    delib_config_category_field = CATEGORY_IA_DELIB_FIELDS_MAPPING_EXTRA_INCLUDE[
+                        self.delib_category_field]
+                    json = response.json()
+                    cat_json = json["extra_include_{categories}".format(
+                        categories=delib_config_category_field)]
 
-                for cat in cat_json:
-                    categories.append((cat['id'], cat['title']))
-                self.delib_categories = categories
-                logger.info("Fetching delib categories for {} [End]".format(self.title))
+                    for cat in cat_json:
+                        categories.append((cat['id'], cat['title']))
+                    self.delib_categories = categories
+                    logger.info("Fetching delib categories for {} [End]".format(self.title))
+                else:
+                    logger.error("Unable to fetch categories, error is {} [End]".format(
+                        response.content))
         return categories