+++ src/plonemeeting/portal/core/tests/test_sync.py
@@ -2,10 +2,10 @@
 from datetime import datetime
 from imio.helpers.content import object_values
 from plone import api
-from plonemeeting.portal.core.browser.sync import get_formatted_data_from_json
-from plonemeeting.portal.core.browser.sync import sync_annexes_data
-from plonemeeting.portal.core.browser.sync import sync_items_data
-from plonemeeting.portal.core.browser.sync import sync_meeting_data
+from plonemeeting.portal.core.sync_utils import get_formatted_data_from_json
+from plonemeeting.portal.core.sync_utils import sync_annexes_data
+from plonemeeting.portal.core.sync_utils import sync_items_data
+from plonemeeting.portal.core.sync_utils import sync_meeting_data
 from plonemeeting.portal.core.content.meeting import IMeeting
 from plonemeeting.portal.core.tests.portal_test_case import PmPortalDemoFunctionalTestCase
 
@@ -66,8 +66,24 @@ class TestMeetingSynchronization(PmPortalDemoFunctionalTestCase):
 
     def test_sync_meeting_items(self):
         meeting = sync_meeting_data(self.institution, self.json_meeting.get("items")[0])
+        # only a few picked items
+        item_external_uids = ['ecd55a85b1ee4039bfe22c7c4988876d',
+                              '12e7d68685074605a2750f0888b0bf52',
+                              '765ec8ae7ec145b987ab9b21ec45ef14',
+                              'aa79bc1b61884e289849999c014acc67']
+        json_items = {"items": [item for item in self.json_meeting_items.get('items')
+                                if item.get('UID') in item_external_uids]
+                      }
+        results = sync_items_data(meeting, json_items, self.institution,
+                                  item_external_uids=item_external_uids + ['fake uid'])
+        self.assertEqual(4, results.get("created"))
+        self.assertListEqual(item_external_uids, [item.plonemeeting_uid for item in meeting.values()])
+        api.content.delete(objects=meeting.values())
+        # all items
         results = sync_items_data(meeting, self.json_meeting_items, self.institution)
-        self.assertEqual(len(meeting.items()), results.get("created"))
+        self.assertEqual(28, results.get("created"))
+        self.assertEqual(0, results.get("modified"))
+        self.assertEqual(0, results.get("deleted"))
         self.assertEqual(meeting.values()[0].number, "1")
         self.assertEqual(meeting.values()[0].sortable_number, 100)
         self.assertEqual(meeting.values()[0].category, 'urbanisme')
@@ -83,7 +99,7 @@ class TestMeetingSynchronization(PmPortalDemoFunctionalTestCase):
         self.assertEqual(meeting.values()[-1].number, "28")
         self.assertEqual(meeting.values()[-1].sortable_number, 2800)
         self.assertEqual(meeting.values()[-1].category, 'locations')
-
+        # force re import everything
         self.institution.delib_category_field = "classifier"
         results = sync_items_data(meeting, self.json_meeting_items, self.institution, True)
         self.assertEqual(len(meeting.items()), results.get("modified"))
@@ -95,11 +111,71 @@ class TestMeetingSynchronization(PmPortalDemoFunctionalTestCase):
 
     def test_sync_with_updates_meeting_items(self):
         meeting = sync_meeting_data(self.institution, self.json_meeting.get("items")[0])
-        # results {'deleted': 0, 'modified': 0, 'created': 28}
-        results = sync_items_data(meeting, self.json_meeting_items, self.institution)
-        self.assertEqual(len(meeting.items()), results.get("created"))
+        # only a few picked items
+        item_external_uids = ['ecd55a85b1ee4039bfe22c7c4988876d',
+                              '12e7d68685074605a2750f0888b0bf52',
+                              '765ec8ae7ec145b987ab9b21ec45ef14',
+                              'aa79bc1b61884e289849999c014acc67']
+        json_items = {"items": [item for item in self.json_meeting_items.get('items')
+                                if item.get('UID') in item_external_uids]
+                      }
+        results = sync_items_data(meeting, json_items, self.institution,
+                                  item_external_uids=item_external_uids + ['fake uid'])
+        self.assertEqual(4, results.get("created"))
+        self.assertEqual(0, results.get("modified"))
+        self.assertEqual(0, results.get("deleted"))
+
+        item_external_uids = ['ecd55a85b1ee4039bfe22c7c4988876d',
+                              'e66269c9342f4e6c861eaff123b20bcb',  # replace
+                              '765ec8ae7ec145b987ab9b21ec45ef14',
+                              'aa79bc1b61884e289849999c014acc67']
+        json_items = {"items": [item for item in self.json_meeting_items.get('items')
+                                if item.get('UID') in item_external_uids]
+                      }
         decision = {"content-type": "text/html", "data": u"<p>Nouvelle décision</p>"}
         modification_date = {"modification_date": u"2019-11-26T14:42:40+00:00"}
+        json_items.get("items")[0].get("decision").update(decision)
+        json_items.get("items")[0].update(modification_date)
+        results = sync_items_data(meeting, json_items, self.institution,
+                                  item_external_uids=item_external_uids + ['fake uid'])
+        self.assertEqual(1, results.get("created"))
+        self.assertEqual(1, results.get("modified"))
+        # the item absent from item_external_uids is not deleted but ignored
+        self.assertEqual(0, results.get("deleted"))
+        self.assertEqual(5, len(meeting.values()))
+        # th 4 item in item_external_uids + the ignored one
+        self.assertListEqual(['ecd55a85b1ee4039bfe22c7c4988876d',
+                              '12e7d68685074605a2750f0888b0bf52',  # ignored
+                              '765ec8ae7ec145b987ab9b21ec45ef14',
+                              'aa79bc1b61884e289849999c014acc67',
+                              'e66269c9342f4e6c861eaff123b20bcb'],  # added
+                             [item.plonemeeting_uid for item in meeting.values()])
+        # one item in the list is not returned -> deleted
+        item_external_uids = ['ecd55a85b1ee4039bfe22c7c4988876d',
+                              '12e7d68685074605a2750f0888b0bf52',  # back but not in json
+                              'e66269c9342f4e6c861eaff123b20bcb',
+                              '765ec8ae7ec145b987ab9b21ec45ef14',
+                              'aa79bc1b61884e289849999c014acc67']
+        results = sync_items_data(meeting, json_items, self.institution,
+                                  item_external_uids=item_external_uids + ['fake uid'])
+        self.assertEqual(0, results.get("created"))
+        self.assertEqual(0, results.get("modified"))
+        # the item present from item_external_uids is not deleted but absent from json is deleted
+        self.assertEqual(1, results.get("deleted"))
+        self.assertEqual(4, len(meeting.values()))
+        self.assertListEqual(['ecd55a85b1ee4039bfe22c7c4988876d',
+                              '765ec8ae7ec145b987ab9b21ec45ef14',
+                              'aa79bc1b61884e289849999c014acc67',
+                              'e66269c9342f4e6c861eaff123b20bcb'],
+                             [item.plonemeeting_uid for item in meeting.values()])
+        # all items
+        # results {'deleted': 0, 'modified': 0, 'created': 28}
+        results = sync_items_data(meeting, self.json_meeting_items, self.institution)
+        self.assertEqual(24, results.get("created"))
+        self.assertEqual(0, results.get("modified"))
+        self.assertEqual(0, results.get("deleted"))
+        decision = {"content-type": "text/html", "data": u"<p>Nouvelle décision</p>"}
+        modification_date = {"modification_date": u"2019-11-26T14:43:40+00:00"}
         self.json_meeting_items.get("items")[0].get("decision").update(decision)
         self.json_meeting_items.get("items")[0].update(modification_date)
         results = sync_items_data(meeting, self.json_meeting_items, self.institution)